Working with assistive gym

+++ a2c_ppo_acktr/algo/gail.py
@@ -78,12 +78,12 @@ class Discriminator(nn.Module):
 
         for expert_batch, policy_batch in zip(expert_loader,
                                               policy_data_generator):
-                                              
+
             policy_state, policy_action = policy_batch[0], policy_batch[3]
             policy_d = self.trunk(torch.cat([policy_state, policy_action], dim=1))
 
             expert_state, expert_action, expert_id = expert_batch
-            expert_state = torch.FloatTensor(expert_state).to(self.device)
+            expert_state = expert_state.to(self.device)
             expert_action = expert_action.to(self.device)
             expert_id = expert_id.to(self.device)
